# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Backend VisionGuard",
        "description": "endpoints del backend",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/admin/pedidos": {
            "get": {
                "tags": [
                    "Pedidos"
                ],
                "summary": "Listar todos los pedidos con los datos de los usuarios",
                "operationId": "6bb5fb43903b305ebced094894cad8d8",
                "responses": {
                    "200": {
                        "description": "Lista de pedidos obtenida correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "pedidos": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id_pedido": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "total": {
                                                        "type": "number",
                                                        "format": "float",
                                                        "example": 150.5
                                                    },
                                                    "usuario": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "integer",
                                                                "example": 1
                                                            },
                                                            "name": {
                                                                "type": "string",
                                                                "example": "John Doe"
                                                            },
                                                            "email": {
                                                                "type": "string",
                                                                "example": "john.doe@example.com"
                                                            },
                                                            "role_id": {
                                                                "type": "integer",
                                                                "example": 2
                                                            },
                                                            "role_name": {
                                                                "type": "string",
                                                                "example": "Cliente"
                                                            },
                                                            "created_at": {
                                                                "type": "string",
                                                                "format": "date-time",
                                                                "example": "2024-01-01T12:00:00.000000Z"
                                                            },
                                                            "updated_at": {
                                                                "type": "string",
                                                                "format": "date-time",
                                                                "example": "2024-01-01T12:00:00.000000Z"
                                                            }
                                                        },
                                                        "type": "object"
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": "2024-01-01T12:00:00.000000Z"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": "2024-01-01T12:00:00.000000Z"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Prohibido"
                    }
                }
            }
        },
...
    },
...
}